feat: 헥사고날 아키텍쳐 기본 구조 설정
