+++ src/main/kotlin/com/ggne/royalflushserver/subway/adapter/in/rest/SubwayController.kt
@@ -1,0 +1,9 @@
+package com.ggne.royalflushserver.subway.adapter.`in`.rest
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/api/more")
+class SubwayController {
+}
