--- src/main/kotlin/com/ggne/royalflushserver/subway/application/DummySubwayRepository.kt
@@ -1,4 +1,0 @@
-package com.ggne.royalflushserver.subway.application
-
-class DummySubwayRepository {
-}
