--- src/main/kotlin/com/ggne/royalflushserver/subway/application/DummySubwayService.kt
@@ -1,4 +1,0 @@
-package com.ggne.royalflushserver.subway.application
-
-class DummySubwayService {
-}
