--- src/main/kotlin/com/ggne/royalflushserver/subway/presentation/SubwayController.kt
@@ -1,9 +1,0 @@
-package com.ggne.royalflushserver.subway.presentation
-
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/api/more")
-class SubwayController {
-}
