+++ src/main/kotlin/com/ggne/royalflushserver/teatime/adapter/in/rest/TeatimeController.kt
@@ -1,0 +1,9 @@
+package com.ggne.royalflushserver.teatime.adapter.`in`.rest
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RequestMapping("/api/teatime")
+@RestController
+class TeatimeController {
+}
