--- src/main/kotlin/com/ggne/royalflushserver/teatime/application/DummyTeatimeRepository.kt
@@ -1,4 +1,0 @@
-package com.ggne.royalflushserver.teatime.application
-
-class DummyTeatimeRepository {
-}
