--- src/main/kotlin/com/ggne/royalflushserver/teatime/presentation/TeatimeController.kt
@@ -1,9 +1,0 @@
-package com.ggne.royalflushserver.teatime.presentation
-
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RequestMapping("/api/teatime")
-@RestController
-class TeatimeController {
-}
