+++ src/main/kotlin/com/ggne/royalflushserver/toilet/adapter/in/rest/ToiletController.kt
@@ -1,0 +1,9 @@
+package com.ggne.royalflushserver.toilet.adapter.`in`.rest
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RequestMapping("/api/more")
+@RestController
+class ToiletController {
+}
