--- src/main/kotlin/com/ggne/royalflushserver/toilet/application/DummyToiletRepository.kt
@@ -1,4 +1,0 @@
-package com.ggne.royalflushserver.toilet.application
-
-class DummyToiletRepository {
-}
