+++ src/main/kotlin/com/ggne/royalflushserver/toilet/domain/vo/Evaluation.kt
@@ -1,0 +1,5 @@
+package com.ggne.royalflushserver.toilet.domain.vo
+
+enum class Evaluation {
+    BAD, OK, GOOD
+}
