--- src/main/kotlin/com/ggne/royalflushserver/toilet/presentation/ToiletController.kt
@@ -1,9 +1,0 @@
-package com.ggne.royalflushserver.toilet.presentation
-
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RequestMapping("/api/more")
-@RestController
-class ToiletController {
-}
