+++ src/main/kotlin/com/ggne/royalflushserver/user/domain/model/User.kt
@@ -1,7 +1,8 @@
 package com.ggne.royalflushserver.user.domain.model
 
 data class User(
-    val id: String,
+    val id: Long? = null,
+    val userId: String,
     val password: String,
     val name: String,
     val coin: Int,
